refactor: Split up routes and add React dep
