+++ client/generate-assets/index.tsx
@@ -1,0 +1,26 @@
+import React from "react";
+
+const IndexPage = () => (
+  <ul>
+    <li>
+      <a href="/badges">Generate badges</a>
+    </li>
+    <li>
+      <a href="/schedule">Generate a schedule poster</a>
+    </li>
+    <li>
+      <a href="/presentation#0">Generate a presentation with speakers</a>
+    </li>
+    <li>
+      <a href="/speakers-room">Generate a speakers' room poster</a>
+    </li>
+    <li>
+      <a href="/left-arrow">Generate a left arrow poster</a>
+    </li>
+    <li>
+      <a href="/right-arrow">Generate a right arrow poster</a>
+    </li>
+  </ul>
+);
+
+export default IndexPage;
