+++ package.json
@@ -26,6 +26,8 @@
     "@types/lodash": "^4.14.120",
     "@types/morgan": "^1.7.35",
     "@types/node": "^10.12.18",
+    "@types/react": "^16.8.2",
+    "@types/react-dom": "^16.8.0",
     "@types/request": "^2.48.1",
     "camelcase": "^5.0.0",
     "flatmap": "0.0.3",
@@ -63,6 +65,8 @@
     "lodash": "^4.17.11",
     "morgan": "^1.9.1",
     "npm": "^6.6.0",
+    "react": "^16.8.1",
+    "react-dom": "^16.8.1",
     "reflect-metadata": "^0.1.13",
     "request": "^2.88.0",
     "rimraf": "^2.6.3",
