+++ server/routes/resolve-image.ts
